\name{coverage}
\alias{coverage}
\docType{data}
\title{
Coverage data of 154 amplicons in 184 breast carcinomas
}
\description{
A matrix of sequencing coverages from semiconductor sequencing
}
\usage{data(coverage)}
\format{
Matrix containing the sequencing coverages of 154 samples (breast carcinomas)
and 184 variables (amplicons).
}

\examples{
data(coverage)
}
\keyword{datasets}
