\name{ioncopy-package}
\alias{ioncopy-package}
\alias{ioncopy}
\docType{package}
\title{
Calling copy number alterations in amplicon sequencing data
}
\description{
Method for the calculation of copy numbers and calling of copy number alterations. The algorithm uses coverage data from amplicon sequencing of a sample cohort as input. The method includes significance assessment, correction for multiple testing and does not depend on normal DNA controls.
}
\details{
\tabular{ll}{
Package: \tab ioncopy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-25\cr
License: \tab GPL-3 \cr
}
}
\author{
Jan Budczies <jan.budczies@charite.de>
}
\references{
Jan Budczies, Nicole Pfarr, Albrecht Stenzinger, Denise Treue, Volker Endris, Fakher Ismaeel, Nikola Bangemann, Jens-Uwe Blohmer, Manfred Dietel, Sibylle Loibl, Wilko Weichert, Carsten Denkert: \emph{Ioncopy: A novel method for calling copy number alterations in amplicon sequencing data including significance assessment.} Submitted.
}
