\name{summarize.CNA}
\alias{summarize.CNA}
\title{
Summary of CNA calls
}
\description{
CNA calls are summarized to tables of all amplifications in each sample and to all amplifications in each amplicon/gene.
}
\usage{
summarize.CNA(calls)
}
\arguments{
  \item{calls}{
List of CNA calls generated by \code{\link{call.CNA}}.
}
}
\value{
List of summary matrices for samples (\code{"samples"}) and amplicons (\code{"amplicon"}) or genes \code{"gene"}.

}
