\name{ioncopy-package}
\alias{ioncopy-package}
\alias{ioncopy}
\docType{package}
\title{
Calling copy number alterations in amplicon sequencing data
}
\description{
Method for the calculation of copy numbers and calling of copy number alterations. The algorithm uses coverage data from amplicon sequencing of a sample cohort as input. The method includes significance assessment, correction for multiple testing and does not depend on normal DNA controls.
}
\details{
\tabular{ll}{
Package: \tab ioncopy\cr
Type: \tab Package\cr
Version: \tab 2.1.1\cr
Date: \tab 2018-03-21\cr
License: \tab GPL-3 \cr
}
}

\author{
Maintainer: Jan Budczies <jan.budczies@charite.de>\cr
Eva Romanovsky <eva.romanovsky@rambler.ru>
}
\references{
Jan Budczies, Nicole Pfarr, Albrecht Stenzinger, Denise Treue, Volker Endris, Fakher Ismaeel, Nikola Bangemann, Jens-Uwe Blohmer, Manfred Dietel, Sibylle Loibl, Wilko Weichert, Carsten Denkert: \emph{Ioncopy: a novel method for calling copy number alterations in amplicon sequencing data including significance assessment.} Oncotarget 7(11):13236-47, 2016, doi: 10.18632/oncotarget.7451.

Jan Budczies, Nicole Pfarr, Eva Romanovsky, Volker Endris, Albrecht Stenzinger, and Carsten Denkert: \emph{Ioncopy: an R Shiny app to call copy number alterations in targeted NGS data.} Submitted.
}
